+++ libraries/apollo-engine-ktor/src/commonMain/kotlin/com/apollographql/apollo3/network/KtorExtensions.kt
@@ -1,6 +1,7 @@
 package com.apollographql.apollo3.network
 
 import com.apollographql.apollo3.ApolloClient
+import com.apollographql.apollo3.annotations.ApolloDeprecatedSince
 import com.apollographql.apollo3.annotations.ApolloExperimental
 import com.apollographql.apollo3.network.http.KtorHttpEngine
 import com.apollographql.apollo3.network.ws.KtorWebSocketEngine
@@ -13,6 +14,8 @@ import io.ktor.client.HttpClient
  * See also [ApolloClient.Builder.httpEngine] and [ApolloClient.Builder.webSocketEngine]
  */
 @ApolloExperimental
+@ApolloDeprecatedSince(ApolloDeprecatedSince.Version.v4_0_0)
+@Deprecated("apollo-engine-ktor has moved to 'com.apollographql.ktor:apollo-engine-ktor'")
 fun ApolloClient.Builder.ktorClient(httpClient: HttpClient) = apply {
   httpEngine(KtorHttpEngine(httpClient))
   webSocketEngine(KtorWebSocketEngine(httpClient))
