+++ libraries/apollo-engine-tests/README.md
@@ -1,0 +1,9 @@
+# Module apollo-engine-tests
+
+apollo-engine-tests contains a list of tests for implementations of `HttpEngine` and `WebSocketEngine`.
+
+This should ideally be apollo-runtime test fixtures.
+
+See https://youtrack.jetbrains.com/issue/KT-46899/Dynamic-test-API
+See https://youtrack.jetbrains.com/issue/KT-63142
+
