+++ libraries/apollo-engine-tests/gradle.properties
@@ -1,0 +1,2 @@
+POM_NAME=Apollo Engine Tests
+POM_DESCRIPTION=Tests for Apollo engines
