+++ libraries/apollo-engine-tests/src/appleMain/kotlin/com/apollographql/apollo3/engine/tests/WebSocketEngineTest.apple.kt
@@ -1,0 +1,5 @@
+package com.apollographql.apollo3.engine.tests
+
+internal actual fun platform(): Platform {
+  return Platform.Native
+}
