+++ libraries/apollo-engine-tests/src/commonMain/kotlin/com/apollographql/apollo3/engine/tests/all-tests.kt
@@ -1,0 +1,19 @@
+package com.apollographql.apollo3.engine.tests
+
+import com.apollographql.apollo3.network.http.HttpEngine
+import com.apollographql.apollo3.network.ws.WebSocketEngine
+
+suspend fun runAllTests(engine: (Long) -> HttpEngine, webSocketEngine: () -> WebSocketEngine, checkCloseFrame: Boolean) {
+  gzipTest(engine(60_000))
+  errorWithBody(engine)
+  headers(engine)
+  post(engine)
+  connectTimeout(engine)
+  readTimeout(engine)
+
+  textFrames(webSocketEngine)
+  binaryFrames(webSocketEngine)
+  serverCloseNicely(webSocketEngine, checkCloseFrame)
+  serverCloseAbruptly(webSocketEngine)
+  headers(webSocketEngine)
+}
